Add remember me functionality for stateless apps
Make personal access tokens valid for 2 hours by default and
long-lived (1 year) by passing remember field on request
and caching the token.

When checking if the token is valid, we also check if there
is a token in the cache, if so, it means that it is long-lived,
so the token will still be valid.